Added some UnityActions
Added some sprites.
Added a basic tower.
Enemies dying runt the action from DeathPortObject and MoneyPortObject.
MoneyPortObject then increases the money on MoneyCounterObject which runs an action that tells the text to update in MoneyTextBehaviour
The action in DeathPortObject is subscribed by WaveHandler that removes the popped enemy from the enemylist.